Ordonnancement tache

+++ src/graphe/Node.java
@@ -26,9 +26,9 @@ public class Node implements Comparable {
     private Float distance=Float.MAX_VALUE;
     private int x=0,y=0;
     private Map<Node, GraphValeur> adjacentNodes = new HashMap<>();
-    public long dateDebutTot=0;
-    public long dateTard=0;
-    public long dureeTache=0;
+    public Float dateDebutTot=0F;
+    public Float dateTard=0F;
+    public Float dureeTache=0F;
     public boolean visited;
     private Timestamp dateDebut;
     private Timestamp dateFin;
@@ -106,7 +106,7 @@ public class Node implements Comparable {
     public void addDestination(Node destination, GraphValeur valeur) {
         adjacentNodes.put(destination, valeur);
         destination.getPredecesseurNoeud().add(this);
-        this.dureeTache=valeur.getDistance().longValue();
+        this.dureeTache=valeur.getDistance();
     }
 
     public Node(String name) {
